gattlib-py: Fix 'gattlib_adapter_scan_*()' prototype after 'timeout' type change

+++ gattlib-py/gattlib/__init__.py
@@ -77,14 +77,14 @@ gattlib_discovered_device_with_data_type = CFUNCTYPE(None, c_void_p, c_char_p, c
                                                      py_object)
 
 # int gattlib_adapter_scan_enable_with_filter(void *adapter, uuid_t **uuid_list, int16_t rssi_threshold, uint32_t enabled_filters,
-#        gattlib_discovered_device_t discovered_device_cb, int timeout, void *user_data)
+#        gattlib_discovered_device_t discovered_device_cb, size_t timeout, void *user_data)
 gattlib_adapter_scan_enable_with_filter = gattlib.gattlib_adapter_scan_enable_with_filter
-gattlib_adapter_scan_enable_with_filter.argtypes = [c_void_p, POINTER(POINTER(GattlibUuid)), c_int16, c_uint32, gattlib_discovered_device_type, c_int, py_object]
+gattlib_adapter_scan_enable_with_filter.argtypes = [c_void_p, POINTER(POINTER(GattlibUuid)), c_int16, c_uint32, gattlib_discovered_device_type, c_size_t, py_object]
 
 # int gattlib_adapter_scan_eddystone(void *adapter, int16_t rssi_threshold, uint32_t eddsytone_types,
-#        gattlib_discovered_device_with_data_t discovered_device_cb, int timeout, void *user_data)
+#        gattlib_discovered_device_with_data_t discovered_device_cb, size_t timeout, void *user_data)
 gattlib_adapter_scan_eddystone = gattlib.gattlib_adapter_scan_eddystone
-gattlib_adapter_scan_eddystone.argtypes = [c_void_p, c_int16, c_uint32, gattlib_discovered_device_with_data_type, c_int, py_object]
+gattlib_adapter_scan_eddystone.argtypes = [c_void_p, c_int16, c_uint32, gattlib_discovered_device_with_data_type, c_size_t, py_object]
 
 # gatt_connection_t *gattlib_connect(const char *src, const char *dst, unsigned long options);
 gattlib_connect = gattlib.gattlib_connect
